Finished code for lab 5b
Created new '/tvshows' route for user to search for a show by calling the TVMaze API, as well as an 'episodes' EJS file to render a list of episodes for a searched show.

Also created an '/api/books' route that returns a list of all books or those matching a search parameter as a JSON object.